+++ ts/packages/core/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @composio/core
 
+## 0.3.1
+
+### Patch Changes
+
+- 73db5f5: Fix callback url not working in toolrouter's session.authorize()
+
+## 0.3.0
+
+### Minor Changes
+
+- 07551cd: Add support for native tool execution in tool router
+- 9e002c5: Minor fixes
+- f0e67c4: Update API client and tool router types
+- 31521bd: Update typedocs and examples for toolkit versions
+- 9e002c5: Alpha release of tool router
+
+## 0.2.7-alpha.4
+
+### Patch Changes
+
+- 07551cd: Add support for native tool execution in tool router
+
+## 0.2.7-alpha.3
+
+### Patch Changes
+
+- f0e67c4: Update API client and tool router types
+
+## 0.2.7-alpha.2
+
+### Patch Changes
+
+- 31521bd: Update typedocs and examples for toolkit versions
+
+## 0.2.7-alpha.1
+
+### Patch Changes
+
+- Minor fixes
+
+## 0.2.7-alpha.0
+
+### Patch Changes
+
+- Alpha release of tool router
+
 ## 0.2.6
 
 ### Patch Changes